+++ src/react-components/src/components/01-atoms/Images/Icons/Icon.tsx
@@ -39,7 +39,7 @@ import bem from "../../../../utils/bem";
             {desc &&
               <desc id={`title-${name}`}>{ desc }</desc>
             }
-            <use xlinkHref={`${Icons}#${name}`} />
+            <use href={`${Icons}#${name}`} />
         </svg>
       );
     }