+++ storybook/.storybook-html/config.js
@@ -2,8 +2,10 @@ import '!style-loader!css-loader!sass-loader!import-glob-loader!design-system-st
 import { configure, load, addDecorator } from '@storybook/html';
 const twig = require('twig');
 import bem from 'design-system-twig/_twig-components/functions/bem';
+import attach_library from 'design-system-twig/_twig-components/functions/sb_attach-library'
 
 twig.extendFunction("bem", bem);
+twig.extendFunction("attach_library", attach_library);
 
 // automatically import all stories from node_modules/stories ending in *.stories.js
 // Direct link used instead of require(module) because of Webpack's inability to resolve variables in require.context