+++ src/react-components/CHANGELOG.md
@@ -12,7 +12,7 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 - `Breadcrumb` now shows icon in mobile view
 
 ### Added
-- `Icon` component that uses Icons.
+- `Icon` component that uses the `design-system-icons` package.
 
 ## [0.0.3] - 2019-11-15
 ### Changed